refactor(app): split app.rs into mod/setup/transcript/render modules
app.rs had grown to ~6.4k lines mixing the event loop, the setup overlay
state machine, runtime-event translation, and all rendering. Split it into
a directory module without changing behavior:

- app/mod.rs       core types + App, the event loop, input/command dispatch
- app/setup.rs     impl App setup/state-machine methods (SetupStep transitions)
- app/transcript.rs runtime event -> ChatLine / status translation
- app/render.rs    all ratatui rendering (chrome, overlay, markdown/diff)

Type definitions stay in mod.rs so the submodules (descendant modules) can
access their fields; moved fns/methods are pub(crate) and re-exported so
existing `crate::app::*` references and the test module resolve unchanged.
SetupStep is documented as the overlay's state machine, with transitions in
setup.rs and rendering in render.rs. The render module is named `render`
(not `draw`) to avoid colliding with the free `draw` fn it exports.

No behavior change: all 335 tests pass.
