+++ src/app/mod.rs
@@ -27,6 +27,16 @@ use std::time::Duration;
 use tokio::sync::broadcast;
 use tokio::sync::mpsc;
 
+mod render;
+mod setup;
+mod transcript;
+
+// Re-export the moved free items so the rest of the crate (and the test module)
+// can keep referring to them as `crate::app::*`. `setup` exposes only `impl App`
+// methods, so it needs no re-export. The rendering module is named `render`
+// rather than `draw` so it does not collide with the free `draw` fn it exports.
+pub(crate) use self::{render::*, transcript::*};
+
 #[derive(Clone, Debug)]
 pub enum Author {
     User,
@@ -141,13 +151,16 @@ pub struct App {
 
 /// Result of one background models API fetch. `Ok(None)` means the provider
 /// does not support listing; the picker keeps the curated fallback list.
-struct ModelDiscovery {
+pub(crate) struct ModelDiscovery {
     provider: String,
     result: Result<Option<Vec<ModelOption>>, String>,
 }
 
+/// State of the first-run / `/setup` overlay. This enum *is* the overlay's
+/// state machine; its transitions (key handling, provider/model discovery,
+/// persistence) live in [`setup`]. Rendering lives in [`render`].
 #[derive(Clone, Debug)]
-enum SetupStep {
+pub(crate) enum SetupStep {
     Provider {
         selected: usize,
     },
@@ -178,7 +191,7 @@ enum SetupStep {
     },
 }
 
-struct ProviderOption {
+pub(crate) struct ProviderOption {
     name: &'static str,
     label: &'static str,
     hint: &'static str,
@@ -222,14 +235,14 @@ const PROVIDER_OPTIONS: &[ProviderOption] = &[
     },
 ];
 
-struct CredentialOption {
+pub(crate) struct CredentialOption {
     id: CredentialAction,
     label: String,
     hint: String,
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum CredentialAction {
+pub(crate) enum CredentialAction {
     UseEnv,
     PasteKey,
     Skip,
@@ -237,13 +250,13 @@ enum CredentialAction {
 }
 
 #[derive(Clone, Debug)]
-struct ModelOption {
+pub(crate) struct ModelOption {
     spec: Option<String>,
     label: String,
     hint: String,
 }
 
-struct EffortOption {
+pub(crate) struct EffortOption {
     value: &'static str,
     label: &'static str,
     hint: &'static str,
@@ -938,1046 +951,6 @@ impl App {
         }
     }
 
-    fn start_setup(&mut self) {
-        self.setup = Some(SetupStep::Provider {
-            selected: self.current_provider_index(),
-        });
-    }
-
-    fn start_first_run_setup(&mut self) {
-        self.start_setup();
-    }
-
-    fn start_model_setup(&mut self) {
-        let provider = self.current_provider_name();
-        self.open_model_step(&provider);
-    }
-
-    fn start_model_setup_with_arg(&mut self, raw: &str) {
-        let spec = raw.trim();
-        if spec.is_empty() {
-            self.start_model_setup();
-            return;
-        }
-        let provider = self.current_provider_name();
-        self.request_model_discovery(&provider);
-        let options = self.model_options(&provider);
-        let selected = model_index_for_label(spec, &options);
-        let custom = if options
-            .get(selected)
-            .and_then(|option| option.spec.as_deref())
-            == Some(spec)
-        {
-            None
-        } else {
-            Some(spec.to_string())
-        };
-        self.setup = Some(SetupStep::PickModel {
-            provider,
-            selected,
-            custom,
-            error: None,
-        });
-    }
-
-    fn start_effort_setup(&mut self, raw: &str) {
-        let raw = raw.trim();
-        let selected = effort_index(raw).unwrap_or_else(|| self.current_effort_index());
-        self.setup = Some(SetupStep::PickEffort {
-            selected,
-            error: if raw.is_empty() || effort_index(raw).is_some() {
-                None
-            } else {
-                Some(format!("unknown effort: {raw}"))
-            },
-        });
-    }
-
-    fn current_effort_index(&self) -> usize {
-        let label = self.model.provider_label();
-        if !label.starts_with("openai/")
-            && !label.starts_with("openrouter/")
-            && !label.starts_with("custom/")
-        {
-            return effort_index("medium").unwrap_or(0);
-        }
-        label
-            .split_whitespace()
-            .nth(1)
-            .and_then(effort_index)
-            .unwrap_or_else(|| effort_index("medium").unwrap_or(0))
-    }
-
-    fn current_provider_name(&self) -> String {
-        self.model
-            .provider_label()
-            .split('/')
-            .next()
-            .unwrap_or("openai")
-            .trim()
-            .to_string()
-    }
-
-    fn current_provider_index(&self) -> usize {
-        let provider = self.current_provider_name();
-        PROVIDER_OPTIONS
-            .iter()
-            .position(|option| option.name == provider)
-            .unwrap_or(0)
-    }
-
-    fn provider_label(name: &str) -> &'static str {
-        PROVIDER_OPTIONS
-            .iter()
-            .find(|option| option.name == name)
-            .map(|option| option.label)
-            .unwrap_or("provider")
-    }
-
-    fn provider_env_names(provider: &str) -> &'static [&'static str] {
-        match provider {
-            "openai" => &["OPENAI_API_KEY"],
-            "anthropic" => &["ANTHROPIC_API_KEY"],
-            "google" => &["GEMINI_API_KEY", "GOOGLE_API_KEY"],
-            "openrouter" => &["OPENROUTER_API_KEY"],
-            "ollama" => &["OLLAMA_BASE_URL", "OLLAMA_API_KEY"],
-            "custom" => &["CUSTOM_API_KEY"],
-            _ => &[],
-        }
-    }
-
-    /// Connection status shown in the provider picker. `true` means the
-    /// provider can be used right away (setup may jump straight to model
-    /// selection); the string is the short status appended to the row hint.
-    fn provider_status(settings: &crate::settings::Settings, provider: &str) -> (bool, String) {
-        match provider {
-            "llmsim" | "ollama" => (true, "✓ no key needed".to_string()),
-            "custom" => {
-                if crate::runtime::custom_base_url(settings).is_some() {
-                    (true, "✓ base URL set".to_string())
-                } else {
-                    (false, "needs base URL".to_string())
-                }
-            }
-            _ => {
-                if let Some(env) = Self::detected_env_var(provider) {
-                    (true, format!("✓ {env}"))
-                } else if settings.has_token(provider) {
-                    (true, "✓ saved key".to_string())
-                } else {
-                    (false, "needs API key".to_string())
-                }
-            }
-        }
-    }
-
-    fn detected_env_var(provider: &str) -> Option<&'static str> {
-        Self::provider_env_names(provider)
-            .iter()
-            .copied()
-            .find(|name| {
-                std::env::var(name)
-                    .map(|value| !value.is_empty())
-                    .unwrap_or(false)
-            })
-    }
-
-    fn credential_options(provider: &str) -> Vec<CredentialOption> {
-        let env_names = Self::provider_env_names(provider);
-        // The custom endpoint treats a key as optional: most local servers
-        // accept any bearer token, so the first option proceeds with the env
-        // key when present and a placeholder otherwise.
-        let env_label = if provider == "custom" {
-            "Continue without key".to_string()
-        } else {
-            match env_names {
-                [] => "Use environment".to_string(),
-                [one] => format!("Use {one} from environment"),
-                many => format!("Use {} from environment", many.join(" / ")),
-            }
-        };
-        let env_hint = Self::detected_env_var(provider)
-            .map(|name| format!("{name} detected"))
-            .unwrap_or_else(|| {
-                if provider == "custom" {
-                    "fine for endpoints without auth".to_string()
-                } else {
-                    "not detected yet".to_string()
-                }
-            });
-        vec![
-            CredentialOption {
-                id: CredentialAction::UseEnv,
-                label: env_label,
-                hint: env_hint,
-            },
-            CredentialOption {
-                id: CredentialAction::PasteKey,
-                label: "Paste API key".to_string(),
-                hint: "saved to settings.toml".to_string(),
-            },
-            CredentialOption {
-                id: CredentialAction::Skip,
-                label: "Skip for now".to_string(),
-                hint: "leave setup unchanged".to_string(),
-            },
-            CredentialOption {
-                id: CredentialAction::ClearSaved,
-                label: "Clear saved key".to_string(),
-                hint: "remove this provider token".to_string(),
-            },
-        ]
-    }
-
-    /// Options for the model picker: models discovered live from the
-    /// provider's models API when available, otherwise the curated
-    /// fallback list.
-    fn model_options(&self, provider: &str) -> Vec<ModelOption> {
-        match self.model_catalog.get(provider) {
-            Some(options) => options.clone(),
-            None => Self::fallback_model_options(provider),
-        }
-    }
-
-    /// Curated static list, used until (or instead of) live discovery —
-    /// e.g. before the models API responds, when the provider does not
-    /// support listing, or when the query fails.
-    fn fallback_model_options(provider: &str) -> Vec<ModelOption> {
-        let mut models = match provider {
-            "openai" => vec![
-                ModelOption {
-                    spec: Some("gpt-5.5".to_string()),
-                    label: "gpt-5.5".to_string(),
-                    hint: "frontier model for complex coding".to_string(),
-                },
-                ModelOption {
-                    spec: Some("gpt-5.4".to_string()),
-                    label: "gpt-5.4".to_string(),
-                    hint: "strong everyday model".to_string(),
-                },
-                ModelOption {
-                    spec: Some("gpt-5.4-mini".to_string()),
-                    label: "gpt-5.4-mini".to_string(),
-                    hint: "fast and cost-efficient".to_string(),
-                },
-                ModelOption {
-                    spec: Some("gpt-5.3-codex".to_string()),
-                    label: "gpt-5.3-codex".to_string(),
-                    hint: "coding-optimized model".to_string(),
-                },
-                ModelOption {
-                    spec: Some("gpt-5.2".to_string()),
-                    label: "gpt-5.2".to_string(),
-                    hint: "optimized for long-running agents".to_string(),
-                },
-            ],
-            "anthropic" => vec![
-                ModelOption {
-                    spec: Some("claude-sonnet-4-5".to_string()),
-                    label: "claude-sonnet-4-5".to_string(),
-                    hint: "best default Claude model".to_string(),
-                },
-                ModelOption {
-                    spec: Some("claude-opus-4-5".to_string()),
-                    label: "claude-opus-4-5".to_string(),
-                    hint: "more capable for complex work".to_string(),
-                },
-                ModelOption {
-                    spec: Some("claude-haiku-4-5".to_string()),
-                    label: "claude-haiku-4-5".to_string(),
-                    hint: "fast answers".to_string(),
-                },
-                ModelOption {
-                    spec: Some("claude-sonnet-4-6".to_string()),
-                    label: "claude-sonnet-4-6".to_string(),
-                    hint: "newer Sonnet option".to_string(),
-                },
-                ModelOption {
-                    spec: Some("claude-fable-5".to_string()),
-                    label: "claude-fable-5".to_string(),
-                    hint: "most powerful Claude model".to_string(),
-                },
-            ],
-            "google" => vec![
-                ModelOption {
-                    spec: Some("gemini-2.5-flash".to_string()),
-                    label: "gemini-2.5-flash".to_string(),
-                    hint: "fast Gemini default".to_string(),
-                },
-                ModelOption {
-                    spec: Some("gemini-2.5-pro".to_string()),
-                    label: "gemini-2.5-pro".to_string(),
-                    hint: "more capable Gemini model".to_string(),
-                },
-            ],
-            "openrouter" => vec![
-                ModelOption {
-                    spec: Some("openai/gpt-5.2".to_string()),
-                    label: "openai/gpt-5.2".to_string(),
-                    hint: "default OpenRouter model".to_string(),
-                },
-                ModelOption {
-                    spec: Some("nvidia/nemotron-3-super-120b-a12b high".to_string()),
-                    label: "nvidia/nemotron-3-super-120b-a12b".to_string(),
-                    hint: "reasoning model through OpenRouter".to_string(),
-                },
-                ModelOption {
-                    spec: Some("anthropic/claude-sonnet-4-5".to_string()),
-                    label: "anthropic/claude-sonnet-4-5".to_string(),
-                    hint: "Claude through OpenRouter".to_string(),
-                },
-            ],
-            "ollama" => vec![ModelOption {
-                spec: Some("llama3.2".to_string()),
-                label: "llama3.2".to_string(),
-                hint: "local default model".to_string(),
-            }],
-            // No preset list exists for an arbitrary endpoint; only the
-            // trailing "Custom..." free-form entry is offered.
-            "custom" => Vec::new(),
-            _ => vec![ModelOption {
-                spec: Some("llmsim-yolop".to_string()),
-                label: "llmsim-yolop".to_string(),
-                hint: "offline demo model".to_string(),
-            }],
-        };
-        models.push(custom_model_option());
-        models
-    }
-
-    fn is_fetching_models(&self, provider: &str) -> bool {
-        self.model_fetches_in_flight.contains(provider)
-    }
-
-    /// Kick off a background fetch of the provider's models API, if one is
-    /// not already cached or in flight. The result lands on `models_rx`
-    /// and is applied between frames; the picker shows the fallback list
-    /// (plus a loading hint) in the meantime.
-    fn request_model_discovery(&mut self, provider: &str) {
-        if !self.model_discovery_enabled
-            || provider == "llmsim"
-            || self.model_catalog.contains_key(provider)
-            || self.model_fetches_in_flight.contains(provider)
-        {
-            return;
-        }
-        // Prefer the live provider choice (it carries any custom base URL);
-        // fall back to the provider's defaults when picking across providers.
-        let choice = if self.current_provider_name() == provider {
-            self.model.provider_choice()
-        } else {
-            match crate::runtime::ProviderChoice::default_for_provider_name(provider) {
-                Ok(choice) => choice,
-                Err(_) => return,
-            }
-        };
-        self.model_fetches_in_flight.insert(provider.to_string());
-        let tx = self.models_tx.clone();
-        let settings = self.settings.clone();
-        let provider_name = provider.to_string();
-        tokio::spawn(async move {
-            let result = match tokio::time::timeout(
-                Duration::from_secs(10),
-                crate::capabilities::model_discovery::discover_provider_models(
-                    &choice,
-                    &settings.snapshot(),
-                ),
-            )
-            .await
-            {
-                Ok(Ok(Some(models))) => Ok(Some(model_options_from_discovered(models))),
-                Ok(Ok(None)) => Ok(None),
-                Ok(Err(err)) => Err(err.to_string()),
-                Err(_) => Err("models API request timed out".to_string()),
-            };
-            let _ = tx.send(ModelDiscovery {
-                provider: provider_name,
-                result,
-            });
-        });
-    }
-
-    /// Apply a finished models API fetch: cache the options and re-anchor
-    /// an open picker for that provider on the refreshed list.
-    fn apply_model_discovery(&mut self, discovery: ModelDiscovery) {
-        self.model_fetches_in_flight.remove(&discovery.provider);
-        match discovery.result {
-            // > 1 because the list always ends with the "Custom..." entry.
-            Ok(Some(options)) if options.len() > 1 => {
-                self.model_catalog
-                    .insert(discovery.provider.clone(), options);
-                if let Some(SetupStep::PickModel {
-                    provider,
-                    custom,
-                    error,
-                    ..
-                }) = self.setup.clone()
-                    && provider == discovery.provider
-                    && custom.is_none()
-                {
-                    let selected = model_index_for_label(
-                        &self.model.model_id(),
-                        &self.model_options(&provider),
-                    );
-                    self.setup = Some(SetupStep::PickModel {
-                        provider,
-                        selected,
-                        custom,
-                        error,
-                    });
-                }
-            }
-            // Listing unsupported (or empty): cache the curated fallback so
-            // reopening the picker doesn't re-query an API that can't answer.
-            // Errors are deliberately not cached — the next picker open retries.
-            Ok(_) => {
-                self.model_catalog.insert(
-                    discovery.provider.clone(),
-                    Self::fallback_model_options(&discovery.provider),
-                );
-            }
-            Err(mut err) => {
-                if let Some(SetupStep::PickModel {
-                    provider, error, ..
-                }) = self.setup.as_mut()
-                    && *provider == discovery.provider
-                {
-                    err.truncate(120);
-                    *error = Some(format!("model list unavailable: {err}"));
-                }
-            }
-        }
-    }
-
-    async fn handle_setup_key(&mut self, key: KeyEvent) {
-        let Some(step) = self.setup.clone() else {
-            return;
-        };
-        match step {
-            SetupStep::Provider { selected } => {
-                self.handle_provider_key(key, selected).await;
-            }
-            SetupStep::BaseUrlInput { value, .. } => {
-                self.handle_base_url_key(key, value).await;
-            }
-            SetupStep::Credential {
-                provider, selected, ..
-            } => {
-                self.handle_credential_key(key, provider, selected).await;
-            }
-            SetupStep::TokenInput {
-                provider, token, ..
-            } => {
-                self.handle_token_key(key, provider, token).await;
-            }
-            SetupStep::PickModel {
-                provider,
-                selected,
-                custom,
-                ..
-            } => {
-                self.handle_model_key(key, provider, selected, custom).await;
-            }
-            SetupStep::PickEffort { selected, .. } => {
-                self.handle_effort_key(key, selected).await;
-            }
-        }
-    }
-
-    async fn handle_provider_key(&mut self, key: KeyEvent, selected: usize) {
-        match key.code {
-            KeyCode::Esc => {
-                self.setup = None;
-            }
-            KeyCode::Up | KeyCode::Char('k') => {
-                self.setup = Some(SetupStep::Provider {
-                    selected: selected.saturating_sub(1),
-                });
-            }
-            KeyCode::Down | KeyCode::Char('j') => {
-                self.setup = Some(SetupStep::Provider {
-                    selected: (selected + 1).min(PROVIDER_OPTIONS.len().saturating_sub(1)),
-                });
-            }
-            KeyCode::Char('c') => {
-                self.open_provider_config(selected);
-            }
-            KeyCode::Char(ch) if ch.is_ascii_digit() => {
-                if let Some(index) = digit_index(ch, PROVIDER_OPTIONS.len()) {
-                    self.confirm_provider(index).await;
-                }
-            }
-            KeyCode::Enter => {
-                self.confirm_provider(selected).await;
-            }
-            _ => {}
-        }
-    }
-
-    /// Enter on a provider row. Already-connected providers switch
-    /// immediately and jump straight to model selection; the rest go
-    /// through credential (or base URL) configuration first.
-    async fn confirm_provider(&mut self, selected: usize) {
-        let option = PROVIDER_OPTIONS
-            .get(selected)
-            .unwrap_or(&PROVIDER_OPTIONS[0]);
-        if option.name == "llmsim" {
-            match self.run_setup_command(Some("provider llmsim")).await {
-                Ok(()) => {
-                    self.setup = None;
-                    self.push_system("setup complete: offline demo mode".into());
-                }
-                Err(error) => {
-                    self.setup = Some(SetupStep::Provider { selected });
-                    self.push_system(format!("setup failed: {error}"));
-                }
-            }
-            return;
-        }
-
-        let (connected, _) = Self::provider_status(&self.settings.snapshot(), option.name);
-        if !connected {
-            self.open_provider_config(selected);
-            return;
-        }
-
-        if option.name == "custom" {
-            // Best-effort switch: it fails harmlessly when no model is saved
-            // yet, and the model step right after sets provider + model
-            // atomically via `provider custom <model>`.
-            let _ = self.run_setup_command(Some("provider custom")).await;
-            self.open_model_step("custom");
-            return;
-        }
-
-        match self
-            .run_setup_command(Some(&format!("provider {}", option.name)))
-            .await
-        {
-            Ok(()) => self.open_model_step(option.name),
-            Err(error) => {
-                // A connected-looking provider that still fails to switch
-                // (stale key, unreachable endpoint) lands on the credential
-                // step where the user can fix it.
-                self.setup = Some(SetupStep::Credential {
-                    provider: option.name.to_string(),
-                    selected: 0,
-                    error: Some(error),
-                });
-            }
-        }
-    }
-
-    /// `c` on a provider row: configure credentials (or the endpoint base
-    /// URL for the custom provider) even when already connected — e.g. to
-    /// replace or clear a saved key.
-    fn open_provider_config(&mut self, selected: usize) {
-        let option = PROVIDER_OPTIONS
-            .get(selected)
-            .unwrap_or(&PROVIDER_OPTIONS[0]);
-        match option.name {
-            "llmsim" => {}
-            "custom" => self.open_base_url_step(),
-            _ => {
-                self.setup = Some(SetupStep::Credential {
-                    provider: option.name.to_string(),
-                    selected: 0,
-                    error: None,
-                });
-            }
-        }
-    }
-
-    fn open_base_url_step(&mut self) {
-        let value = crate::runtime::custom_base_url(&self.settings.snapshot()).unwrap_or_default();
-        self.setup = Some(SetupStep::BaseUrlInput { value, error: None });
-    }
-
-    /// Open the model picker for `provider`, preselecting the active model
-    /// and kicking off live discovery. The custom provider has no curated
-    /// list, so until (or unless) discovery fills the catalog its picker
-    /// shows only the free-form "Custom..." entry.
-    fn open_model_step(&mut self, provider: &str) {
-        self.request_model_discovery(provider);
-        let selected = model_index_for_label(&self.model.model_id(), &self.model_options(provider));
-        self.setup = Some(SetupStep::PickModel {
-            provider: provider.to_string(),
-            selected,
-            custom: None,
-            error: None,
-        });
-    }
-
-    /// Prefill for the custom provider's free-form model input: the active
-    /// model when already on the custom provider, else the saved one.
-    fn custom_model_prefill(&self) -> String {
-        if self.current_provider_name() == "custom" {
-            return self.model.model_id();
-        }
-        self.settings
-            .snapshot()
-            .model_for("custom")
-            .unwrap_or_default()
-            .to_string()
-    }
-
-    async fn handle_base_url_key(&mut self, key: KeyEvent, mut value: String) {
-        match key.code {
-            KeyCode::Esc => {
-                self.setup = Some(SetupStep::Provider {
-                    selected: PROVIDER_OPTIONS
-                        .iter()
-                        .position(|option| option.name == "custom")
-                        .unwrap_or(0),
-                });
-            }
-            KeyCode::Enter => {
-                let trimmed = value.trim().to_string();
-                if trimmed.is_empty() {
-                    self.setup = Some(SetupStep::BaseUrlInput {
-                        value,
-                        error: Some(
-                            "enter a base URL like http://localhost:8000/v1, or press Esc"
-                                .to_string(),
-                        ),
-                    });
-                    return;
-                }
-                match self
-                    .run_setup_command(Some(&format!("url custom {trimmed}")))
-                    .await
-                {
-                    Ok(()) => {
-                        self.setup = Some(SetupStep::Credential {
-                            provider: "custom".to_string(),
-                            selected: 0,
-                            error: None,
-                        });
-                    }
-                    Err(error) => {
-                        self.setup = Some(SetupStep::BaseUrlInput {
-                            value,
-                            error: Some(error),
-                        });
-                    }
-                }
-            }
-            KeyCode::Backspace => {
-                value.pop();
-                self.setup = Some(SetupStep::BaseUrlInput { value, error: None });
-            }
-            KeyCode::Char(_) => {
-                if let KeyCode::Char(ch) = normalize_printable_key(key).code {
-                    value.push(ch);
-                }
-                self.setup = Some(SetupStep::BaseUrlInput { value, error: None });
-            }
-            _ => {}
-        }
-    }
-
-    async fn handle_credential_key(&mut self, key: KeyEvent, provider: String, selected: usize) {
-        let options = Self::credential_options(&provider);
-        match key.code {
-            KeyCode::Esc => {
-                self.setup = Some(SetupStep::Provider {
-                    selected: PROVIDER_OPTIONS
-                        .iter()
-                        .position(|option| option.name == provider)
-                        .unwrap_or(0),
-                });
-            }
-            KeyCode::Up | KeyCode::Char('k') => {
-                self.setup = Some(SetupStep::Credential {
-                    provider,
-                    selected: selected.saturating_sub(1),
-                    error: None,
-                });
-            }
-            KeyCode::Down | KeyCode::Char('j') => {
-                self.setup = Some(SetupStep::Credential {
-                    provider,
-                    selected: (selected + 1).min(options.len().saturating_sub(1)),
-                    error: None,
-                });
-            }
-            KeyCode::Char(ch) if ch.is_ascii_digit() => {
-                if let Some(index) = digit_index(ch, options.len()) {
-                    self.confirm_credential(provider, index).await;
-                }
-            }
-            KeyCode::Enter => {
-                self.confirm_credential(provider, selected).await;
-            }
-            _ => {}
-        }
-    }
-
-    async fn confirm_credential(&mut self, provider: String, selected: usize) {
-        let options = Self::credential_options(&provider);
-        let action = options
-            .get(selected)
-            .map(|option| option.id)
-            .unwrap_or(CredentialAction::UseEnv);
-        match action {
-            CredentialAction::UseEnv => {
-                // The custom provider can't be switched to until a model is
-                // chosen, so the validation switch is deferred to the model
-                // step (its key is optional anyway).
-                if provider == "custom" {
-                    self.open_model_step("custom");
-                    return;
-                }
-                match self
-                    .run_setup_command(Some(&format!("provider {provider}")))
-                    .await
-                {
-                    Ok(()) => self.open_model_step(&provider),
-                    Err(error) => {
-                        self.setup = Some(SetupStep::Credential {
-                            provider,
-                            selected,
-                            error: Some(error),
-                        });
-                    }
-                }
-            }
-            CredentialAction::PasteKey => {
-                self.setup = Some(SetupStep::TokenInput {
-                    provider,
-                    token: String::new(),
-                    error: None,
-                });
-            }
-            CredentialAction::Skip => {
-                self.setup = None;
-                self.push_system("setup skipped".into());
-            }
-            CredentialAction::ClearSaved => {
-                let result = self
-                    .run_setup_command(Some(&format!("token {provider} clear")))
-                    .await;
-                match result {
-                    Ok(()) => {
-                        self.setup = None;
-                        self.push_system(format!(
-                            "setup complete: cleared saved key for {provider}"
-                        ));
-                    }
-                    Err(error) => {
-                        self.setup = Some(SetupStep::Credential {
-                            provider,
-                            selected,
-                            error: Some(error),
-                        });
-                    }
-                }
-            }
-        }
-    }
-
-    async fn handle_token_key(&mut self, key: KeyEvent, provider: String, mut token: String) {
-        match key.code {
-            KeyCode::Esc => {
-                self.setup = Some(SetupStep::Credential {
-                    provider,
-                    selected: 1,
-                    error: None,
-                });
-            }
-            KeyCode::Enter => {
-                let trimmed = token.trim().to_string();
-                if trimmed.is_empty() {
-                    self.setup = Some(SetupStep::TokenInput {
-                        provider,
-                        token,
-                        error: Some("API key is empty — paste a key, or press Esc".to_string()),
-                    });
-                    return;
-                }
-                let save_arg = format!("token {provider} {trimmed}");
-                if let Err(error) = self.run_setup_command(Some(&save_arg)).await {
-                    self.setup = Some(SetupStep::TokenInput {
-                        provider,
-                        token: String::new(),
-                        error: Some(error),
-                    });
-                    return;
-                }
-                // Custom defers the provider switch to the model step (no
-                // model is known yet); other providers validate the new key
-                // by switching now.
-                if provider == "custom" {
-                    self.open_model_step("custom");
-                    return;
-                }
-                match self
-                    .run_setup_command(Some(&format!("provider {provider}")))
-                    .await
-                {
-                    Ok(()) => self.open_model_step(&provider),
-                    Err(error) => {
-                        self.setup = Some(SetupStep::TokenInput {
-                            provider,
-                            token: String::new(),
-                            error: Some(error),
-                        });
-                    }
-                }
-            }
-            KeyCode::Backspace => {
-                token.pop();
-                self.setup = Some(SetupStep::TokenInput {
-                    provider,
-                    token,
-                    error: None,
-                });
-            }
-            KeyCode::Char(_) => {
-                if let KeyCode::Char(ch) = normalize_printable_key(key).code {
-                    token.push(ch);
-                }
-                self.setup = Some(SetupStep::TokenInput {
-                    provider,
-                    token,
-                    error: None,
-                });
-            }
-            _ => {}
-        }
-    }
-
-    async fn handle_model_key(
-        &mut self,
-        key: KeyEvent,
-        provider: String,
-        selected: usize,
-        custom: Option<String>,
-    ) {
-        let options = self.model_options(&provider);
-        if let Some(mut value) = custom {
-            match key.code {
-                KeyCode::Esc => {
-                    self.setup = Some(SetupStep::PickModel {
-                        provider,
-                        selected,
-                        custom: None,
-                        error: None,
-                    });
-                }
-                KeyCode::Enter => {
-                    let value = value.trim().to_string();
-                    if value.is_empty() {
-                        self.setup = Some(SetupStep::PickModel {
-                            provider,
-                            selected,
-                            custom: Some(String::new()),
-                            error: Some("paste a model id, or press Esc to go back".to_string()),
-                        });
-                        return;
-                    }
-                    self.save_model_and_finish(&provider, &value, selected)
-                        .await;
-                }
-                KeyCode::Backspace => {
-                    value.pop();
-                    self.setup = Some(SetupStep::PickModel {
-                        provider,
-                        selected,
-                        custom: Some(value),
-                        error: None,
-                    });
-                }
-                KeyCode::Char(_) => {
-                    if let KeyCode::Char(ch) = normalize_printable_key(key).code {
-                        value.push(ch);
-                    }
-                    self.setup = Some(SetupStep::PickModel {
-                        provider,
-                        selected,
-                        custom: Some(value),
-                        error: None,
-                    });
-                }
-                _ => {}
-            }
-            return;
-        }
-
-        match key.code {
-            KeyCode::Esc => {
-                self.setup = Some(SetupStep::Provider {
-                    selected: PROVIDER_OPTIONS
-                        .iter()
-                        .position(|option| option.name == provider)
-                        .unwrap_or(0),
-                });
-            }
-            KeyCode::Up | KeyCode::Char('k') => {
-                self.setup = Some(SetupStep::PickModel {
-                    provider,
-                    selected: selected.saturating_sub(1),
-                    custom: None,
-                    error: None,
-                });
-            }
-            KeyCode::Down | KeyCode::Char('j') => {
-                self.setup = Some(SetupStep::PickModel {
-                    provider,
-                    selected: (selected + 1).min(options.len().saturating_sub(1)),
-                    custom: None,
-                    error: None,
-                });
-            }
-            KeyCode::Char(ch) if ch.is_ascii_digit() => {
-                if let Some(index) = digit_index(ch, options.len()) {
-                    self.confirm_model(provider, index).await;
-                }
-            }
-            KeyCode::Enter => {
-                self.confirm_model(provider, selected).await;
-            }
-            _ => {}
-        }
-    }
-
-    async fn confirm_model(&mut self, provider: String, selected: usize) {
-        let options = self.model_options(&provider);
-        let Some(option) = options.get(selected) else {
-            self.setup = Some(SetupStep::PickModel {
-                provider,
-                selected: 0,
-                custom: None,
-                error: None,
-            });
-            return;
-        };
-        if let Some(spec) = &option.spec {
-            self.save_model_and_finish(&provider, spec, selected).await;
-        } else {
-            // "Custom..." free-form input. For the custom endpoint the
-            // current/saved model is prefilled — it is the provider's only
-            // memory of a model id.
-            let prefill = if provider == "custom" {
-                self.custom_model_prefill()
-            } else {
-                String::new()
-            };
-            self.setup = Some(SetupStep::PickModel {
-                provider,
-                selected,
-                custom: Some(prefill),
-                error: None,
-            });
-        }
-    }
-
-    async fn save_model_and_finish(&mut self, provider: &str, spec: &str, selected: usize) {
-        // `model <spec>` resolves against the current provider. For the
-        // custom endpoint the wizard reaches this step before any switch
-        // could succeed (a first-time custom provider has no model yet), so
-        // it switches provider and model atomically instead.
-        let command = if provider == "custom" {
-            format!("provider custom {spec}")
-        } else {
-            format!("model {spec}")
-        };
-        match self.run_setup_command(Some(&command)).await {
-            Ok(()) => {
-                self.setup = None;
-                self.push_system(format!("setup complete: {}", self.model.provider_label()));
-            }
-            Err(error) => {
-                self.setup = Some(SetupStep::PickModel {
-                    provider: provider.to_string(),
-                    selected,
-                    custom: None,
-                    error: Some(error),
-                });
-            }
-        }
-    }
-
-    async fn handle_effort_key(&mut self, key: KeyEvent, selected: usize) {
-        match key.code {
-            KeyCode::Esc => {
-                self.setup = None;
-            }
-            KeyCode::Up | KeyCode::Char('k') => {
-                self.setup = Some(SetupStep::PickEffort {
-                    selected: selected.saturating_sub(1),
-                    error: None,
-                });
-            }
-            KeyCode::Down | KeyCode::Char('j') => {
-                self.setup = Some(SetupStep::PickEffort {
-                    selected: (selected + 1).min(EFFORT_OPTIONS.len().saturating_sub(1)),
-                    error: None,
-                });
-            }
-            KeyCode::Char(ch) if ch.is_ascii_digit() => {
-                if let Some(index) = digit_index(ch, EFFORT_OPTIONS.len()) {
-                    self.confirm_effort(index).await;
-                }
-            }
-            KeyCode::Enter => {
-                self.confirm_effort(selected).await;
-            }
-            _ => {}
-        }
-    }
-
-    async fn confirm_effort(&mut self, selected: usize) {
-        let Some(option) = EFFORT_OPTIONS.get(selected) else {
-            self.setup = Some(SetupStep::PickEffort {
-                selected: 0,
-                error: None,
-            });
-            return;
-        };
-        match self
-            .run_setup_command(Some(&format!("effort {}", option.value)))
-            .await
-        {
-            Ok(()) => {
-                self.setup = None;
-                self.push_system(format!("setup complete: {}", self.model.provider_label()));
-            }
-            Err(error) => {
-                self.setup = Some(SetupStep::PickEffort {
-                    selected,
-                    error: Some(error),
-                });
-            }
-        }
-    }
-
-    async fn execute_setup_command(
-        &mut self,
-        arg: Option<&str>,
-    ) -> Result<everruns_core::command::CommandResult, String> {
-        let request = everruns_core::command::ExecuteCommandRequest {
-            name: "setup".to_string(),
-            arguments: arg.map(str::to_string),
-            controls: None,
-        };
-        self.handles
-            .runtime
-            .execute_command(self.handles.session_id, request)
-            .await
-            .map_err(|err| err.to_string())
-    }
-
-    /// Execute an internal setup command. Successful mutations stay quiet so
-    /// the user only sees the overlay and a single final completion line.
-    async fn run_setup_command(&mut self, arg: Option<&str>) -> Result<(), String> {
-        match self.execute_setup_command(arg).await {
-            Ok(result) if result.success => Ok(()),
-            Ok(result) => Err(result.message),
-            Err(err) => Err(err),
-        }
-    }
-
     fn start_turn(&mut self, prompt: String) {
         let handles = self.handles.clone();
         let model = self.model.clone();
@@ -2126,281 +1099,6 @@ pub(crate) struct DeltaRouter {
     last_tool_call: Option<String>,
 }
 
-pub(crate) fn handle_live_event(
-    event: &RuntimeEvent,
-    emitted_events: &mut HashSet<String>,
-    router: &mut DeltaRouter,
-    tx: &mpsc::UnboundedSender<TurnEvent>,
-) {
-    if !emitted_events.insert(event.id.to_string()) {
-        return;
-    }
-
-    match &event.data {
-        EventData::OutputMessageDelta(data) => {
-            router.last_assistant_turn = Some(data.turn_id);
-            let _ = tx.send(TurnEvent::Stream(Some(StreamPreview {
-                kind: StreamKind::Assistant,
-                text: data.accumulated.clone(),
-            })));
-            return;
-        }
-        EventData::OutputMessageCompleted(_) | EventData::OutputMessageReplaced(_)
-            if router.last_assistant_turn.is_some() =>
-        {
-            router.last_assistant_turn = None;
-            let _ = tx.send(TurnEvent::Stream(None));
-        }
-        EventData::ReasonThinkingDelta(data) => {
-            router.last_thinking_turn = Some(data.turn_id);
-            let _ = tx.send(TurnEvent::Stream(Some(StreamPreview {
-                kind: StreamKind::Thinking,
-                text: data.accumulated.clone(),
-            })));
-            return;
-        }
-        EventData::ReasonThinkingCompleted(_) if router.last_thinking_turn.is_some() => {
-            router.last_thinking_turn = None;
-            let _ = tx.send(TurnEvent::Stream(None));
-        }
-        EventData::ToolOutputDelta(data) => {
-            router.last_tool_call = Some(data.tool_call_id.clone());
-            let text = format!(
-                "{} [{}] {}",
-                data.tool_name,
-                data.stream,
-                data.delta.trim_end()
-            );
-            let _ = tx.send(TurnEvent::Stream(Some(StreamPreview {
-                kind: StreamKind::Tool,
-                text,
-            })));
-            return;
-        }
-        EventData::ToolCompleted(data)
-            if router.last_tool_call.as_deref() == Some(data.tool_call_id.as_str()) =>
-        {
-            router.last_tool_call = None;
-            let _ = tx.send(TurnEvent::Stream(None));
-        }
-        _ => {}
-    }
-
-    if let Some(activity) = status_for_event(event) {
-        let _ = tx.send(TurnEvent::Activity(activity));
-    }
-    let lines = lines_for_event(event);
-    if !lines.is_empty() {
-        let _ = tx.send(TurnEvent::Lines(lines));
-    }
-}
-
-/// Drain any persisted events (from `runtime.events()`) that the broadcast
-/// receiver may have missed — used after a `Lagged` recv error and once
-/// more at end-of-turn so the transcript is never missing tool/reason
-/// completion lines.
-async fn catch_up_events(
-    handles: &RuntimeHandles,
-    events_before: usize,
-    emitted_events: &mut HashSet<String>,
-    tx: &mpsc::UnboundedSender<TurnEvent>,
-) {
-    let events = handles.runtime.events().await.unwrap_or_default();
-    let mut lines = Vec::new();
-    for event in events.iter().skip(events_before) {
-        let event_id = event.id.to_string();
-        if !emitted_events.insert(event_id) {
-            continue;
-        }
-        if let Some(activity) = status_for_event(event) {
-            let _ = tx.send(TurnEvent::Activity(activity));
-        }
-        lines.extend(lines_for_event(event));
-    }
-    if !lines.is_empty() {
-        let _ = tx.send(TurnEvent::Lines(lines));
-    }
-}
-
-pub fn lines_for_event(event: &RuntimeEvent) -> Vec<ChatLine> {
-    match &event.data {
-        EventData::ReasonStarted(_) => Vec::new(),
-        EventData::ReasonCompleted(data) => {
-            // Render pre-tool-call narration. When there are no tool calls
-            // the final assistant message will arrive via the message loop
-            // shortly, so we'd duplicate it; keep the has_tool_calls gate.
-            if data.success && data.has_tool_calls {
-                let mut lines = Vec::new();
-                if let Some(text) = data
-                    .text_preview
-                    .as_deref()
-                    .map(str::trim)
-                    .filter(|text| !text.is_empty())
-                {
-                    lines.push(ChatLine {
-                        author: Author::Narration,
-                        text: text.to_string(),
-                    });
-                }
-                lines
-            } else {
-                Vec::new()
-            }
-        }
-        EventData::ReasonItem(data) => data
-            .summary
-            .iter()
-            .filter_map(|segment| {
-                let trimmed = segment.trim();
-                (!trimmed.is_empty()).then(|| ChatLine {
-                    author: Author::Narration,
-                    text: trimmed.to_string(),
-                })
-            })
-            .collect(),
-        EventData::OutputMessageCompleted(_) => Vec::new(),
-        EventData::ToolCompleted(data) => {
-            if data.tool_name == "write_todos" {
-                return todo_lines_for_result(data);
-            }
-            let marker = if data.success { "✓" } else { "✗" };
-            let label = data
-                .narration
-                .as_deref()
-                .or(data.display_name.as_deref())
-                .unwrap_or(data.tool_name.as_str());
-            let summary = summarize_tool_result(data);
-            let mut lines = vec![ChatLine {
-                author: Author::Tool,
-                text: if summary.is_empty() {
-                    format!("{marker} {label}")
-                } else {
-                    format!("{marker} {label}  {summary}")
-                },
-            }];
-            if data.tool_name == "edit_file"
-                && let Some(diff) = extract_field(data, "diff")
-            {
-                for line in diff.lines().take(40) {
-                    lines.push(ChatLine {
-                        author: Author::Diff,
-                        text: line.to_string(),
-                    });
-                }
-            }
-            lines
-        }
-        _ => Vec::new(),
-    }
-}
-
-fn lines_for_replayed_event(event: &RuntimeEvent) -> Vec<ChatLine> {
-    match &event.data {
-        EventData::InputMessage(data) => message_line(Author::User, &data.message)
-            .into_iter()
-            .collect(),
-        EventData::OutputMessageCompleted(data) => {
-            if data.message.role == MessageRole::Agent {
-                message_line(Author::Assistant, &data.message)
-                    .into_iter()
-                    .collect()
-            } else {
-                Vec::new()
-            }
-        }
-        _ => lines_for_event(event),
-    }
-}
-
-fn message_line(author: Author, message: &Message) -> Option<ChatLine> {
-    let text = message.text()?;
-    let text = text.trim();
-    if text.is_empty() {
-        return None;
-    }
-    Some(ChatLine {
-        author,
-        text: text.to_string(),
-    })
-}
-
-pub fn status_for_event(event: &RuntimeEvent) -> Option<ActivityStatus> {
-    match &event.data {
-        EventData::ReasonStarted(_) => Some(fallback_status("thinking")),
-        EventData::ReasonCompleted(data) => {
-            if !data.success {
-                let err = data.error.as_deref().unwrap_or("reasoning failed");
-                return Some(activity_status(format!(
-                    "reasoning failed: {}",
-                    first_line(err, 100)
-                )));
-            }
-            data.has_tool_calls
-                .then(|| activity_status(format!("planned {} tool call(s)", data.tool_call_count)))
-        }
-        EventData::ActStarted(data) => data
-            .headline
-            .clone()
-            .or_else(|| Some(format!("running {} tool(s)", data.tool_calls.len())))
-            .map(activity_status),
-        EventData::ActCompleted(data) => data
-            .headline
-            .clone()
-            .or_else(|| {
-                Some(format!(
-                    "tools finished: {} ok, {} failed",
-                    data.success_count, data.error_count
-                ))
-            })
-            .map(activity_status),
-        EventData::ToolStarted(data) => Some(activity_status(format!(
-            "→ {}",
-            data.narration
-                .as_deref()
-                .or(data.display_name.as_deref())
-                .unwrap_or(data.tool_call.name.as_str())
-        ))),
-        EventData::ToolProgress(data) => Some(activity_status(format!(
-            "… {}: {}",
-            data.display_name
-                .as_deref()
-                .unwrap_or(data.tool_name.as_str()),
-            first_line(&data.message, 100)
-        ))),
-        EventData::ToolCallRequested(data) => Some(activity_status(format!(
-            "waiting for {} client tool result(s)",
-            data.tool_calls.len()
-        ))),
-        EventData::OutputMessageStarted(data) => {
-            let iteration = data.iteration.unwrap_or(1);
-            Some(activity_status(format!(
-                "iteration {iteration}: writing response"
-            )))
-        }
-        EventData::ReasonThinkingStarted(_) => Some(fallback_status("thinking deeply")),
-        EventData::TurnCancelled(_) => Some(activity_status("turn cancelled")),
-        EventData::TurnFailed(data) => Some(activity_status(format!(
-            "turn failed: {}",
-            first_line(&data.error, 100)
-        ))),
-        _ => None,
-    }
-}
-
-fn activity_status(text: impl Into<String>) -> ActivityStatus {
-    ActivityStatus {
-        text: text.into(),
-        fallback: false,
-    }
-}
-
-fn fallback_status(text: impl Into<String>) -> ActivityStatus {
-    ActivityStatus {
-        text: text.into(),
-        fallback: true,
-    }
-}
-
 fn command_suggestions(
     input: &str,
     capability_commands: &[CommandDescriptor],
@@ -2541,1462 +1239,6 @@ fn shifted_char(ch: char) -> Option<char> {
         _ => return None,
     };
     Some(shifted)
-}
-
-// ---------- helpers for surfacing tool results ----------
-
-fn result_value(data: &ToolCompletedData) -> Option<Value> {
-    let parts = data.result.as_ref()?;
-    for part in parts {
-        if let ContentPart::Text(t) = part
-            && let Ok(v) = serde_json::from_str::<Value>(&t.text)
-        {
-            return Some(v);
-        }
-    }
-    None
-}
-
-fn extract_field(data: &ToolCompletedData, field: &str) -> Option<String> {
-    let v = result_value(data)?;
-    v.get(field).and_then(|s| s.as_str()).map(str::to_string)
-}
-
-const MAX_RENDERED_TODOS: usize = 20;
-const MAX_TODO_TEXT_CHARS: usize = 160;
-
-fn truncate_chars(value: &str, max_chars: usize) -> String {
-    let mut chars = value.chars();
-    let truncated: String = chars.by_ref().take(max_chars).collect();
-    if chars.next().is_some() {
-        format!("{truncated}…")
-    } else {
-        truncated
-    }
-}
-
-fn todo_lines_for_result(data: &ToolCompletedData) -> Vec<ChatLine> {
-    let Some(v) = result_value(data) else {
-        return vec![ChatLine {
-            author: Author::Tool,
-            text: format!(
-                "✓ {}",
-                data.display_name.as_deref().unwrap_or("Write Todos")
-            ),
-        }];
-    };
-    let Some(todos) = v.get("todos").and_then(Value::as_array) else {
-        return vec![ChatLine {
-            author: Author::Tool,
-            text: summarize_tool_result(data),
-        }];
-    };
-
-    let total = todos.len();
-    let completed = todos
-        .iter()
-        .filter(|todo| todo.get("status").and_then(Value::as_str) == Some("completed"))
-        .count();
-    let summary = format!("{completed} of {total} todos completed");
-    let mut rendered_todos = Vec::new();
-    for todo in todos.iter().take(MAX_RENDERED_TODOS) {
-        let status = todo
-            .get("status")
-            .and_then(Value::as_str)
-            .unwrap_or("pending");
-        let content = todo.get("content").and_then(Value::as_str).unwrap_or("");
-        let active_form = todo
-            .get("activeForm")
-            .and_then(Value::as_str)
-            .unwrap_or(content);
-        let (icon, text) = match status {
-            "completed" => ("✓", content),
-            "in_progress" => ("›", active_form),
-            _ => ("○", content),
-        };
-        rendered_todos.push(format!(
-            "{icon} {}",
-            truncate_chars(text, MAX_TODO_TEXT_CHARS)
-        ));
-    }
-
-    let mut lines = if rendered_todos.len() <= 3 {
-        let inline_todos = rendered_todos.join("  ");
-        vec![ChatLine {
-            author: Author::Tool,
-            text: if inline_todos.is_empty() {
-                summary
-            } else {
-                format!("{summary}  {inline_todos}")
-            },
-        }]
-    } else {
-        let mut lines = vec![ChatLine {
-            author: Author::Tool,
-            text: summary,
-        }];
-        lines.extend(rendered_todos.into_iter().map(|text| ChatLine {
-            author: Author::ToolDetail,
-            text,
-        }));
-        lines
-    };
-
-    let omitted = total.saturating_sub(MAX_RENDERED_TODOS);
-    if omitted > 0 {
-        lines.push(ChatLine {
-            author: Author::ToolDetail,
-            text: format!("… {omitted} more todo(s) omitted"),
-        });
-    }
-
-    if let Some(warning) = v.get("warning").and_then(Value::as_str) {
-        lines.push(ChatLine {
-            author: Author::ToolDetail,
-            text: format!("warning: {}", truncate_chars(warning, MAX_TODO_TEXT_CHARS)),
-        });
-    }
-
-    lines
-}
-
-/// One-line summary of a tool result, used in the transcript and `--print` output.
-pub fn summarize_tool_result(data: &ToolCompletedData) -> String {
-    let Some(v) = result_value(data) else {
-        if let Some(err) = &data.error {
-            return format!("error: {}", first_line(err, 120));
-        }
-        return String::new();
-    };
-    // Field names match the built-in `session_file_system` capability's
-    // result shapes. See crates/core/src/capabilities/file_system.rs.
-    match data.tool_name.as_str() {
-        "write_todos" => {
-            let completed = v.get("completed").and_then(Value::as_u64).unwrap_or(0);
-            let total = v.get("total_tasks").and_then(Value::as_u64).unwrap_or(0);
-            format!("{completed}/{total} completed")
-        }
-        "read_file" => {
-            let path = v.get("path").and_then(Value::as_str).unwrap_or("");
-            let total = v.get("total_lines").and_then(Value::as_u64).unwrap_or(0);
-            let shown = v.get("lines_shown");
-            let start = shown
-                .and_then(|s| s.get("start"))
-                .and_then(Value::as_u64)
-                .unwrap_or(0);
-            let end = shown
-                .and_then(|s| s.get("end"))
-                .and_then(Value::as_u64)
-                .unwrap_or(0);
-            let count = end.saturating_sub(start.saturating_sub(1));
-            format!("{path} ({count}/{total} lines)")
-        }
-        "write_file" => {
-            let path = v.get("path").and_then(Value::as_str).unwrap_or("");
-            let bytes = v.get("size_bytes").and_then(Value::as_u64).unwrap_or(0);
-            format!("{path} ({bytes} bytes)")
-        }
-        "edit_file" => {
-            let path = v.get("path").and_then(Value::as_str).unwrap_or("");
-            let n = v.get("applied_edits").and_then(Value::as_u64).unwrap_or(0);
-            format!("{path} ({n} edit(s))")
-        }
-        "list_directory" => {
-            let path = v.get("path").and_then(Value::as_str).unwrap_or("");
-            let n = v.get("count").and_then(Value::as_u64).unwrap_or(0);
-            format!("{path} ({n} entries)")
-        }
-        "grep_files" => {
-            let pattern = v.get("pattern").and_then(Value::as_str).unwrap_or("");
-            let n = v.get("match_count").and_then(Value::as_u64).unwrap_or(0);
-            format!("/{pattern}/ ({n} match(es))")
-        }
-        "delete_file" => {
-            let path = v.get("path").and_then(Value::as_str).unwrap_or("");
-            format!("{path} (deleted)")
-        }
-        "stat_file" => {
-            let path = v.get("path").and_then(Value::as_str).unwrap_or("");
-            let size = v.get("size_bytes").and_then(Value::as_u64).unwrap_or(0);
-            format!("{path} ({size} bytes)")
-        }
-        "bash" => {
-            let cmd = v
-                .get("command")
-                .and_then(Value::as_str)
-                .map(|c| first_line(c, 80))
-                .unwrap_or_default();
-            let code = v
-                .get("exit_code")
-                .and_then(Value::as_i64)
-                .map(|c| c.to_string())
-                .unwrap_or_else(|| "?".into());
-            format!("`{cmd}` exit={code}")
-        }
-        _ => String::new(),
-    }
-}
-
-fn first_line(s: &str, max: usize) -> String {
-    let l = s.lines().next().unwrap_or("");
-    if l.len() > max {
-        format!("{}…", &l[..max])
-    } else {
-        l.to_string()
-    }
-}
-
-// ---------- rendering ----------
-
-fn draw(f: &mut ratatui::Frame, app: &mut App) {
-    let input_height = app.input_height();
-    let area = f.area();
-    let state = app.view_state();
-    let chrome_area = bottom_rect(area, chrome_height(input_height));
-    let transcript_area = Rect {
-        height: area.height.saturating_sub(chrome_area.height),
-        ..area
-    };
-
-    // Chrome renders the non-input rows; we then layer the input field
-    // on top into the chrome-reserved input slot.
-    clear_transcript_viewport(f, transcript_area);
-    draw_recent_transcript(f, transcript_area, app);
-    let input_rect = draw_chrome(f, chrome_area, input_height, &state);
-    draw_input(f, input_rect, app);
-    draw_setup_overlay(f, area, app);
-}
-
-fn bottom_rect(area: Rect, height: u16) -> Rect {
-    let height = height.min(area.height);
-    Rect {
-        y: area.y + area.height.saturating_sub(height),
-        height,
-        ..area
-    }
-}
-
-fn chrome_height(input_height: u16) -> u16 {
-    COMPACT_CHROME_HEIGHT.max(input_height.saturating_add(2))
-}
-
-fn clear_transcript_viewport(f: &mut ratatui::Frame, area: Rect) {
-    if area.width == 0 || area.height == 0 {
-        return;
-    }
-
-    f.render_widget(Clear, area);
-}
-
-fn draw_recent_transcript(f: &mut ratatui::Frame, area: Rect, app: &App) {
-    if area.width < 4 || area.height == 0 {
-        return;
-    }
-
-    let inner = Rect {
-        x: area.x.saturating_add(1),
-        y: area.y,
-        width: area.width.saturating_sub(2),
-        height: area.height,
-    };
-    if inner.width == 0 || inner.height == 0 {
-        return;
-    }
-
-    let rendered = recent_transcript_lines(app, inner.width as usize, inner.height as usize);
-    if rendered.is_empty() {
-        return;
-    }
-
-    f.render_widget(Paragraph::new(rendered), inner);
-}
-
-fn recent_transcript_lines(app: &App, width: usize, max_lines: usize) -> Vec<Line<'static>> {
-    if max_lines == 0 {
-        return Vec::new();
-    }
-
-    let mut chunks = Vec::new();
-    let mut total_lines = 0;
-    let mut newer_author: Option<Author> = None;
-
-    for chat in app
-        .lines
-        .iter()
-        .rev()
-        .filter(|line| !matches!(line.author, Author::System))
-        .take(RECENT_TRANSCRIPT_SOURCE_LINES)
-    {
-        let chat = bounded_recent_chat_line(chat);
-        let mut chunk = Vec::new();
-        append_chat_lines(&mut chunk, &chat, width);
-        if should_insert_chat_gap(&chat.author, newer_author.as_ref()) {
-            chunk.push(Line::from(""));
-        }
-
-        if total_lines + chunk.len() > max_lines {
-            let remaining = max_lines.saturating_sub(total_lines);
-            if remaining > 0 {
-                chunks.push(chunk.split_off(chunk.len().saturating_sub(remaining)));
-            }
-            break;
-        }
-
-        total_lines += chunk.len();
-        newer_author = Some(chat.author);
-        chunks.push(chunk);
-    }
-
-    chunks.reverse();
-    chunks.into_iter().flatten().collect()
-}
-
-fn bounded_recent_chat_line(chat: &ChatLine) -> ChatLine {
-    if chat.text.len() <= RECENT_TRANSCRIPT_MAX_TEXT_BYTES {
-        return chat.clone();
-    }
-
-    ChatLine {
-        author: chat.author.clone(),
-        text: truncate_tail_bytes(&chat.text, RECENT_TRANSCRIPT_MAX_TEXT_BYTES),
-    }
-}
-
-/// Render the non-input chrome (command suggestions / stream preview,
-/// message separator, status separator, session status) into `area` using `state`, and
-/// return the `Rect` where the caller should render the input widget
-/// (which needs `&mut` and so cannot be driven through `ViewState`).
-///
-/// Snapshot tests call this against a `TestBackend` and ignore the
-/// returned input rect — the buffer's other rows are what they assert
-/// against.
-pub(crate) fn draw_chrome(
-    f: &mut ratatui::Frame,
-    area: Rect,
-    input_height: u16,
-    state: &ViewState,
-) -> Rect {
-    let preview_height = u16::from(input_height == 1);
-    let status_height = u16::from(input_height < 3);
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Length(preview_height), // suggestions / stream preview
-            Constraint::Length(1),              // message separator
-            Constraint::Length(input_height),   // input (left to the caller)
-            Constraint::Length(status_height),  // status separator
-            Constraint::Length(1),              // session status
-        ])
-        .split(area);
-
-    if state.command_suggestions.is_empty() {
-        draw_stream_preview(f, chunks[0], state);
-    } else {
-        draw_suggestions(f, chunks[0], &state.command_suggestions);
-    }
-    draw_message_separator(f, chunks[1], state);
-    draw_status_separator(f, chunks[3]);
-    draw_session_status(f, chunks[4], state);
-
-    chunks[2]
-}
-
-fn draw_setup_overlay(f: &mut ratatui::Frame, area: Rect, app: &App) {
-    if app.setup.is_none() || area.width == 0 || area.height == 0 {
-        return;
-    }
-    let panel = setup_panel_rect(area);
-    if panel.width == 0 || panel.height == 0 {
-        return;
-    }
-    f.render_widget(Clear, panel);
-    let block = Block::default()
-        .borders(Borders::ALL)
-        .style(Style::default().bg(PANEL_BG).fg(TEXT_PRIMARY));
-    f.render_widget(block, panel);
-    let inner = Rect {
-        x: panel.x.saturating_add(2),
-        y: panel.y.saturating_add(1),
-        width: panel.width.saturating_sub(4),
-        height: panel.height.saturating_sub(2),
-    };
-    let (lines, cursor) = setup_overlay_content(app);
-    f.render_widget(
-        Paragraph::new(lines).style(Style::default().bg(PANEL_BG)),
-        inner,
-    );
-    if let Some((row, col)) = cursor
-        && inner.height > 0
-        && inner.width > 0
-    {
-        f.set_cursor_position((
-            inner
-                .x
-                .saturating_add((col as u16).min(inner.width.saturating_sub(1))),
-            inner
-                .y
-                .saturating_add((row as u16).min(inner.height.saturating_sub(1))),
-        ));
-    }
-}
-
-fn setup_panel_rect(area: Rect) -> Rect {
-    let width = area.width.saturating_sub(4).min(104).max(area.width.min(1));
-    let height = area
-        .height
-        .saturating_sub(2)
-        .min(18)
-        .max(area.height.min(1));
-    Rect {
-        x: area.x + area.width.saturating_sub(width) / 2,
-        y: area.y + area.height.saturating_sub(height) / 2,
-        width,
-        height,
-    }
-}
-
-fn setup_overlay_content(app: &App) -> (Vec<Line<'static>>, Option<(usize, usize)>) {
-    let mut lines = Vec::new();
-    let mut cursor = None;
-    match app.setup.as_ref() {
-        Some(SetupStep::Provider { selected }) => {
-            lines.push(setup_title("Set Up Yolop"));
-            lines.push(setup_hint(
-                "Connected providers jump straight to model selection.",
-            ));
-            lines.push(Line::from(""));
-            let current = app.current_provider_name();
-            let snapshot = app.settings.snapshot();
-            for (idx, option) in PROVIDER_OPTIONS.iter().enumerate() {
-                let (_, status) = App::provider_status(&snapshot, option.name);
-                let mut hint = format!("{} · {status}", option.hint);
-                if option.name == current {
-                    hint.push_str(" · current");
-                }
-                lines.push(setup_row(idx == *selected, idx + 1, option.label, &hint));
-            }
-            lines.push(Line::from(""));
-            lines.push(setup_footer(
-                "Enter select · c configure key/URL · ↑/↓ move · Esc cancel",
-            ));
-        }
-        Some(SetupStep::BaseUrlInput { value, error }) => {
-            lines.push(setup_title("Custom OpenAI-Compatible Endpoint"));
-            lines.push(setup_hint(
-                "Base URL of the API, e.g. http://localhost:8000/v1 — saved to settings.toml.",
-            ));
-            lines.push(Line::from(""));
-            let input = format!("› {value}");
-            cursor = Some((3, input.chars().count()));
-            lines.push(Line::from(vec![
-                Span::styled(
-                    "› ",
-                    Style::default()
-                        .fg(ACCENT_BLUE)
-                        .add_modifier(Modifier::BOLD),
-                ),
-                Span::styled(value.clone(), Style::default().fg(TEXT_PRIMARY)),
-            ]));
-            push_setup_error(&mut lines, error.as_deref());
-            lines.push(setup_footer("Enter save · Esc back"));
-        }
-        Some(SetupStep::Credential {
-            provider,
-            selected,
-            error,
-            ..
-        }) => {
-            lines.push(setup_title(&format!(
-                "API Key for {}",
-                App::provider_label(provider)
-            )));
-            lines.push(setup_hint(
-                "Choose how yolop should authenticate this provider.",
-            ));
-            lines.push(Line::from(""));
-            for (idx, option) in App::credential_options(provider).iter().enumerate() {
-                lines.push(setup_row(
-                    idx == *selected,
-                    idx + 1,
-                    &option.label,
-                    &option.hint,
-                ));
-            }
-            push_setup_error(&mut lines, error.as_deref());
-            lines.push(setup_footer("Enter confirm · ↑/↓ move · Esc back"));
-        }
-        Some(SetupStep::TokenInput {
-            provider,
-            token,
-            error,
-            ..
-        }) => {
-            lines.push(setup_title(&format!(
-                "Paste API Key for {}",
-                App::provider_label(provider)
-            )));
-            lines.push(setup_hint(
-                "The key is masked and is never written to the transcript.",
-            ));
-            lines.push(Line::from(""));
-            let masked = if token.is_empty() {
-                String::new()
-            } else {
-                "•".repeat(token.chars().count())
-            };
-            let input = format!("› {masked}");
-            cursor = Some((3, input.chars().count()));
-            lines.push(Line::from(vec![
-                Span::styled(
-                    "› ",
-                    Style::default()
-                        .fg(ACCENT_BLUE)
-                        .add_modifier(Modifier::BOLD),
-                ),
-                Span::styled(masked, Style::default().fg(TEXT_PRIMARY)),
-            ]));
-            push_setup_error(&mut lines, error.as_deref());
-            lines.push(setup_footer("Enter save · Esc back"));
-        }
-        Some(SetupStep::PickModel {
-            provider,
-            selected,
-            custom,
-            error,
-        }) => {
-            lines.push(setup_title("Select Model"));
-            lines.push(setup_hint(&if provider == "custom" {
-                "Model id served by your endpoint. Applies to this session and future sessions."
-                    .to_string()
-            } else {
-                format!(
-                    "{} models. Applies to this session and future sessions.",
-                    App::provider_label(provider)
-                )
-            }));
-            lines.push(Line::from(""));
-            let options = app.model_options(provider);
-            if let Some(value) = custom {
-                let input = format!("› {value}");
-                cursor = Some((3, input.chars().count()));
-                lines.push(Line::from(vec![
-                    Span::styled(
-                        "› ",
-                        Style::default()
-                            .fg(ACCENT_BLUE)
-                            .add_modifier(Modifier::BOLD),
-                    ),
-                    Span::styled(value.clone(), Style::default().fg(TEXT_PRIMARY)),
-                ]));
-            } else {
-                if app.is_fetching_models(provider) {
-                    lines.push(setup_hint("fetching models from the provider API…"));
-                }
-                let total = options.len();
-                let (start, end) = model_window(*selected, total, MAX_VISIBLE_MODEL_ROWS);
-                if start > 0 {
-                    lines.push(setup_hint(&format!("↑ {start} more")));
-                }
-                // Specs are provider-relative; compare against the bare model
-                // id (the same anchor `model_index_for_label` uses).
-                let current = app.model.model_id();
-                for (idx, option) in options.iter().enumerate().take(end).skip(start) {
-                    let mut hint = option.hint.to_string();
-                    if option.spec.as_deref() == Some(current.as_str()) {
-                        hint.push_str(" · current");
-                    }
-                    lines.push(setup_row(idx == *selected, idx + 1, &option.label, &hint));
-                }
-                if end < total {
-                    lines.push(setup_hint(&format!("↓ {} more", total - end)));
-                }
-            }
-            push_setup_error(&mut lines, error.as_deref());
-            lines.push(setup_footer("Enter confirm · ↑/↓ move · Esc back"));
-        }
-        Some(SetupStep::PickEffort { selected, error }) => {
-            lines.push(setup_title("Select Reasoning Effort"));
-            lines.push(setup_hint(
-                "Applies to OpenAI, OpenRouter, and custom endpoints — this session and future sessions.",
-            ));
-            lines.push(Line::from(""));
-            let label = app.model.provider_label();
-            let current = if label.starts_with("openai/") {
-                label
-                    .split_whitespace()
-                    .nth(1)
-                    .unwrap_or("medium")
-                    .to_string()
-            } else if label.starts_with("openrouter/") || label.starts_with("custom/") {
-                label
-                    .split_whitespace()
-                    .nth(1)
-                    .unwrap_or_default()
-                    .to_string()
-            } else {
-                String::new()
-            };
-            for (idx, option) in EFFORT_OPTIONS.iter().enumerate() {
-                let mut hint = option.hint.to_string();
-                if option.value == current {
-                    hint.push_str(" · current");
-                }
-                lines.push(setup_row(idx == *selected, idx + 1, option.label, &hint));
-            }
-            push_setup_error(&mut lines, error.as_deref());
-            lines.push(setup_footer("Enter confirm · ↑/↓ move · Esc cancel"));
-        }
-        None => {}
-    }
-    (lines, cursor)
-}
-
-fn setup_title(text: &str) -> Line<'static> {
-    Line::from(Span::styled(
-        text.to_string(),
-        Style::default()
-            .fg(TEXT_PRIMARY)
-            .bg(PANEL_BG)
-            .add_modifier(Modifier::BOLD),
-    ))
-}
-
-fn setup_hint(text: &str) -> Line<'static> {
-    Line::from(Span::styled(
-        text.to_string(),
-        Style::default().fg(TEXT_MUTED).bg(PANEL_BG),
-    ))
-}
-
-fn setup_footer(text: &str) -> Line<'static> {
-    Line::from(Span::styled(
-        text.to_string(),
-        Style::default().fg(TEXT_MUTED).bg(PANEL_BG),
-    ))
-}
-
-fn setup_row(selected: bool, index: usize, label: &str, hint: &str) -> Line<'static> {
-    let marker = if selected { "›" } else { " " };
-    let marker_style = if selected {
-        Style::default()
-            .fg(ACCENT_BLUE)
-            .bg(PANEL_BG)
-            .add_modifier(Modifier::BOLD)
-    } else {
-        Style::default().fg(TEXT_DIM).bg(PANEL_BG)
-    };
-    let label_style = if selected {
-        Style::default()
-            .fg(Color::Rgb(135, 220, 205))
-            .bg(PANEL_BG)
-            .add_modifier(Modifier::BOLD)
-    } else {
-        Style::default().fg(TEXT_PRIMARY).bg(PANEL_BG)
-    };
-    Line::from(vec![
-        Span::styled(format!("{marker} "), marker_style),
-        Span::styled(
-            format!("{index}. "),
-            Style::default().fg(TEXT_MUTED).bg(PANEL_BG),
-        ),
-        // Pad to a 28-col label column so hints align, but always keep at
-        // least a 2-space gap: labels like "Use OPENAI_API_KEY from
-        // environment" overflow the column, and a bare `{:<28}` would let the
-        // hint butt right against them ("environmentnot detected yet").
-        Span::styled(
-            {
-                let pad = 28usize.saturating_sub(label.chars().count()).max(2);
-                format!("{label}{}", " ".repeat(pad))
-            },
-            label_style,
-        ),
-        Span::styled(
-            hint.to_string(),
-            Style::default().fg(TEXT_MUTED).bg(PANEL_BG),
-        ),
-    ])
-}
-
-fn push_setup_error(lines: &mut Vec<Line<'static>>, error: Option<&str>) {
-    if let Some(error) = error {
-        lines.push(Line::from(""));
-        lines.push(Line::from(Span::styled(
-            format!("error: {error}"),
-            Style::default().fg(Color::Rgb(220, 120, 90)).bg(PANEL_BG),
-        )));
-    } else {
-        lines.push(Line::from(""));
-    }
-}
-
-fn draw_suggestions(f: &mut ratatui::Frame, area: Rect, suggestions: &[CommandSuggestion]) {
-    if area.height == 0 || area.width == 0 {
-        return;
-    }
-    f.render_widget(
-        Paragraph::new(suggestion_preview_line(suggestions, area.width)),
-        area,
-    );
-}
-
-fn suggestion_preview_line(suggestions: &[CommandSuggestion], width: u16) -> Line<'static> {
-    let body = suggestions
-        .iter()
-        .map(|suggestion| suggestion.label.as_str())
-        .collect::<Vec<_>>()
-        .join("  ·  ");
-    let prefix = "Tab ";
-    let max_body = (width as usize)
-        .saturating_sub(prefix.chars().count() + 1)
-        .max(8);
-    Line::from(vec![
-        Span::styled(
-            prefix,
-            Style::default()
-                .fg(ACCENT_BLUE)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::styled(
-            truncate_end_chars(&body, max_body),
-            Style::default().fg(TEXT_MUTED),
-        ),
-    ])
-}
-
-fn draw_stream_preview(f: &mut ratatui::Frame, area: Rect, state: &ViewState) {
-    if area.height == 0 {
-        return;
-    }
-    let Some(preview) = state.stream_preview.as_ref() else {
-        return;
-    };
-    let inner_width = area.width as usize;
-    if inner_width == 0 {
-        return;
-    }
-    // Show the most recent line of the accumulated stream so the eye
-    // tracks the live tail rather than the start of a long response.
-    let tail = preview
-        .text
-        .lines()
-        .rev()
-        .find(|l| !l.trim().is_empty())
-        .unwrap_or("");
-    let label = preview.kind.label();
-    let prefix = format!("{label} › ");
-    let prefix_w = prefix.chars().count();
-    let max_text = inner_width.saturating_sub(prefix_w + 1).max(8);
-    let truncated = truncate_tail_chars(tail, max_text);
-    f.render_widget(
-        Paragraph::new(Line::from(vec![
-            Span::styled(
-                prefix,
-                Style::default()
-                    .fg(preview.kind.color())
-                    .add_modifier(Modifier::DIM),
-            ),
-            Span::styled(truncated, Style::default().fg(TEXT_MUTED)),
-        ])),
-        area,
-    );
-}
-
-/// Keep the last `max_chars` of `text`. Streaming preview reads better
-/// when the cursor (tail of the stream) is what's visible.
-fn truncate_tail_chars(text: &str, max_chars: usize) -> String {
-    let count = text.chars().count();
-    if count <= max_chars {
-        return text.to_string();
-    }
-    let skip = count - max_chars.saturating_sub(1);
-    let mut out = String::with_capacity(max_chars);
-    out.push('…');
-    out.extend(text.chars().skip(skip));
-    out
-}
-
-fn truncate_tail_bytes(text: &str, max_bytes: usize) -> String {
-    if text.len() <= max_bytes {
-        return text.to_string();
-    }
-    if max_bytes == 0 {
-        return String::new();
-    }
-    if max_bytes <= '…'.len_utf8() {
-        return "…".to_string();
-    }
-
-    let mut start = text.len().saturating_sub(max_bytes - '…'.len_utf8());
-    while start < text.len() && !text.is_char_boundary(start) {
-        start += 1;
-    }
-    format!("…{}", &text[start..])
-}
-
-fn truncate_end_chars(text: &str, max_chars: usize) -> String {
-    let count = text.chars().count();
-    if count <= max_chars {
-        return text.to_string();
-    }
-    if max_chars == 0 {
-        return String::new();
-    }
-    if max_chars == 1 {
-        return "…".to_string();
-    }
-    let mut out = String::with_capacity(max_chars);
-    out.extend(text.chars().take(max_chars - 1));
-    out.push('…');
-    out
-}
-
-fn should_insert_chat_gap(current: &Author, next: Option<&Author>) -> bool {
-    let Some(next) = next else {
-        return false;
-    };
-
-    !matches!(
-        (current, next),
-        (&Author::Tool, &Author::Tool)
-            | (&Author::Tool, &Author::ToolDetail)
-            | (&Author::ToolDetail, &Author::Tool)
-            | (&Author::ToolDetail, &Author::ToolDetail)
-    )
-}
-
-fn append_chat_lines<'a>(lines: &mut Vec<Line<'a>>, chat: &ChatLine, inner_width: usize) {
-    if matches!(chat.author, Author::ToolDetail) {
-        append_wrapped_plain(
-            lines,
-            "           ",
-            Style::default().fg(TEXT_MUTED),
-            &chat.text,
-            inner_width,
-        );
-        return;
-    }
-
-    let header_text = format!("{} › ", chat.author.label());
-    let header_style = Style::default()
-        .fg(chat.author.color())
-        .add_modifier(Modifier::BOLD);
-    if matches!(chat.author, Author::Assistant) {
-        append_markdown_lines(lines, &header_text, header_style, &chat.text, inner_width);
-    } else if matches!(chat.author, Author::Narration) {
-        append_wrapped_styled(
-            lines,
-            &header_text,
-            header_style,
-            &chat.text,
-            inner_width,
-            Style::default().fg(TEXT_MUTED),
-        );
-    } else if matches!(chat.author, Author::Diff) {
-        append_wrapped_diff(lines, &header_text, header_style, &chat.text, inner_width);
-    } else {
-        append_wrapped_plain(lines, &header_text, header_style, &chat.text, inner_width);
-    }
-}
-
-fn append_wrapped_plain<'a>(
-    lines: &mut Vec<Line<'a>>,
-    first_prefix: &str,
-    prefix_style: Style,
-    text: &str,
-    inner_width: usize,
-) {
-    append_wrapped_styled(
-        lines,
-        first_prefix,
-        prefix_style,
-        text,
-        inner_width,
-        Style::default(),
-    );
-}
-
-fn append_wrapped_styled<'a>(
-    lines: &mut Vec<Line<'a>>,
-    first_prefix: &str,
-    prefix_style: Style,
-    text: &str,
-    inner_width: usize,
-    content_style: Style,
-) {
-    let continuation = " ".repeat(first_prefix.chars().count());
-    let wrap_width = inner_width
-        .saturating_sub(first_prefix.chars().count())
-        .max(20);
-    let mut emitted = false;
-    for raw in text.lines() {
-        let wrapped = textwrap::wrap(raw, wrap_width);
-        if wrapped.is_empty() {
-            let prefix = if emitted {
-                continuation.as_str()
-            } else {
-                first_prefix
-            };
-            lines.push(Line::from(vec![Span::styled(
-                prefix.to_string(),
-                prefix_style,
-            )]));
-            emitted = true;
-            continue;
-        }
-        for piece in wrapped {
-            let prefix = if emitted {
-                continuation.as_str()
-            } else {
-                first_prefix
-            };
-            lines.push(Line::from(vec![
-                Span::styled(prefix.to_string(), prefix_style),
-                Span::styled(piece.into_owned(), content_style),
-            ]));
-            emitted = true;
-        }
-    }
-    if !emitted {
-        lines.push(Line::from(vec![Span::styled(
-            first_prefix.to_string(),
-            prefix_style,
-        )]));
-    }
-}
-
-fn append_wrapped_diff<'a>(
-    lines: &mut Vec<Line<'a>>,
-    first_prefix: &str,
-    prefix_style: Style,
-    text: &str,
-    inner_width: usize,
-) {
-    let continuation = " ".repeat(first_prefix.chars().count());
-    let wrap_width = inner_width
-        .saturating_sub(first_prefix.chars().count())
-        .max(20);
-    let mut emitted = false;
-    for raw in text.lines() {
-        let content_style = diff_line_style(raw);
-        let wrapped = textwrap::wrap(raw, wrap_width);
-        if wrapped.is_empty() {
-            let prefix = if emitted {
-                continuation.as_str()
-            } else {
-                first_prefix
-            };
-            lines.push(Line::from(vec![Span::styled(
-                prefix.to_string(),
-                prefix_style,
-            )]));
-            emitted = true;
-            continue;
-        }
-        for piece in wrapped {
-            let prefix = if emitted {
-                continuation.as_str()
-            } else {
-                first_prefix
-            };
-            lines.push(Line::from(vec![
-                Span::styled(prefix.to_string(), prefix_style),
-                Span::styled(piece.into_owned(), content_style),
-            ]));
-            emitted = true;
-        }
-    }
-    if !emitted {
-        lines.push(Line::from(vec![Span::styled(
-            first_prefix.to_string(),
-            prefix_style,
-        )]));
-    }
-}
-
-fn diff_line_style(line: &str) -> Style {
-    let color = if line.starts_with('+') {
-        DIFF_ADD
-    } else if line.starts_with('-') {
-        DIFF_DELETE
-    } else if line.starts_with("@@") || line.starts_with('\\') {
-        DIFF_META
-    } else {
-        TEXT_PRIMARY
-    };
-    Style::default().fg(color)
-}
-
-fn append_markdown_lines<'a>(
-    lines: &mut Vec<Line<'a>>,
-    first_prefix: &str,
-    prefix_style: Style,
-    text: &str,
-    inner_width: usize,
-) {
-    let continuation = " ".repeat(first_prefix.chars().count());
-    let wrap_width = inner_width
-        .saturating_sub(first_prefix.chars().count())
-        .max(20);
-    let mut first = true;
-    let mut in_code = false;
-
-    for raw in text.lines() {
-        let trimmed = raw.trim_end();
-        if let Some(lang) = trimmed.trim_start().strip_prefix("```") {
-            in_code = !in_code;
-            let code_lang = lang.trim();
-            let label = if in_code {
-                if code_lang.is_empty() {
-                    "code".to_string()
-                } else {
-                    format!("code: {code_lang}")
-                }
-            } else {
-                String::new()
-            };
-            push_markdown_line(
-                lines,
-                first_prefix,
-                &continuation,
-                prefix_style,
-                &mut first,
-                vec![Span::styled(
-                    label,
-                    Style::default().fg(TEXT_DIM).add_modifier(Modifier::ITALIC),
-                )],
-            );
-            continue;
-        }
-
-        let content_spans = if in_code {
-            markdown_code_spans(trimmed)
-        } else {
-            markdown_text_spans(trimmed)
-        };
-        let plain = spans_plain_text(&content_spans);
-        let wrapped = textwrap::wrap(&plain, wrap_width);
-        if wrapped.is_empty() {
-            push_markdown_line(
-                lines,
-                first_prefix,
-                &continuation,
-                prefix_style,
-                &mut first,
-                vec![],
-            );
-            continue;
-        }
-        if content_spans.len() == 1 {
-            let style = content_spans[0].style;
-            for piece in wrapped {
-                push_markdown_line(
-                    lines,
-                    first_prefix,
-                    &continuation,
-                    prefix_style,
-                    &mut first,
-                    vec![Span::styled(piece.into_owned(), style)],
-                );
-            }
-        } else {
-            push_markdown_line(
-                lines,
-                first_prefix,
-                &continuation,
-                prefix_style,
-                &mut first,
-                content_spans,
-            );
-        }
-    }
-}
-
-fn push_markdown_line<'a>(
-    lines: &mut Vec<Line<'a>>,
-    first_prefix: &str,
-    continuation: &str,
-    prefix_style: Style,
-    first: &mut bool,
-    mut spans: Vec<Span<'a>>,
-) {
-    let prefix = if *first { first_prefix } else { continuation };
-    let mut line_spans = vec![Span::styled(prefix.to_string(), prefix_style)];
-    line_spans.append(&mut spans);
-    lines.push(Line::from(line_spans));
-    *first = false;
-}
-
-fn markdown_text_spans(text: &str) -> Vec<Span<'static>> {
-    let trimmed = text.trim_start();
-    if trimmed.starts_with('#') {
-        let heading = trimmed.trim_start_matches('#').trim_start();
-        return vec![Span::styled(
-            heading.to_string(),
-            Style::default()
-                .fg(TEXT_PRIMARY)
-                .add_modifier(Modifier::BOLD),
-        )];
-    }
-    if let Some(rest) = trimmed.strip_prefix("> ") {
-        return vec![
-            Span::styled("| ", Style::default().fg(ACCENT_BLUE)),
-            Span::styled(rest.to_string(), Style::default().fg(TEXT_MUTED)),
-        ];
-    }
-    if let Some(rest) = trimmed
-        .strip_prefix("- ")
-        .or_else(|| trimmed.strip_prefix("* "))
-    {
-        return vec![
-            Span::styled("- ", Style::default().fg(ACCENT_GOLD)),
-            Span::raw(rest.to_string()),
-        ];
-    }
-    if let Some((marker, rest)) = numbered_marker(trimmed) {
-        return vec![
-            Span::styled(marker, Style::default().fg(ACCENT_GOLD)),
-            Span::raw(rest.to_string()),
-        ];
-    }
-    inline_code_spans(text)
-}
-
-fn markdown_code_spans(text: &str) -> Vec<Span<'static>> {
-    let mut spans = vec![Span::styled("    ", Style::default().fg(TEXT_DIM))];
-    spans.extend(simple_code_highlight(text));
-    spans
-}
-
-fn inline_code_spans(text: &str) -> Vec<Span<'static>> {
-    let mut spans = Vec::new();
-    let mut rest = text;
-    let mut code = false;
-    while let Some((before, after_tick)) = rest.split_once('`') {
-        if !before.is_empty() {
-            spans.push(Span::raw(before.to_string()));
-        }
-        if let Some((inside, after)) = after_tick.split_once('`') {
-            spans.push(Span::styled(
-                inside.to_string(),
-                Style::default().fg(TEXT_PRIMARY).bg(CODE_BG),
-            ));
-            rest = after;
-            code = true;
-        } else {
-            spans.push(Span::raw("`".to_string()));
-            rest = after_tick;
-            break;
-        }
-    }
-    if !rest.is_empty() {
-        spans.push(Span::raw(rest.to_string()));
-    }
-    if spans.is_empty() || !code {
-        vec![Span::raw(text.to_string())]
-    } else {
-        spans
-    }
-}
-
-fn simple_code_highlight(text: &str) -> Vec<Span<'static>> {
-    let keywords = [
-        "async", "await", "const", "enum", "fn", "impl", "let", "match", "pub", "return", "struct",
-        "use",
-    ];
-    let mut spans = Vec::new();
-    let mut token = String::new();
-    for ch in text.chars() {
-        if ch.is_alphanumeric() || ch == '_' {
-            token.push(ch);
-            continue;
-        }
-        if !token.is_empty() {
-            let style = if keywords.contains(&token.as_str()) {
-                Style::default()
-                    .fg(ACCENT_GOLD)
-                    .add_modifier(Modifier::BOLD)
-            } else if token.chars().all(|c| c.is_ascii_digit()) {
-                Style::default().fg(TEXT_MUTED)
-            } else {
-                Style::default().fg(ACCENT_BLUE)
-            };
-            spans.push(Span::styled(std::mem::take(&mut token), style));
-        }
-        spans.push(Span::styled(ch.to_string(), Style::default().fg(TEXT_DIM)));
-    }
-    if !token.is_empty() {
-        let style = if keywords.contains(&token.as_str()) {
-            Style::default()
-                .fg(ACCENT_GOLD)
-                .add_modifier(Modifier::BOLD)
-        } else if token.chars().all(|c| c.is_ascii_digit()) {
-            Style::default().fg(TEXT_MUTED)
-        } else {
-            Style::default().fg(ACCENT_BLUE)
-        };
-        spans.push(Span::styled(token, style));
-    }
-    spans
-}
-
-fn numbered_marker(text: &str) -> Option<(String, &str)> {
-    let dot = text.find(". ")?;
-    if dot == 0 || !text[..dot].chars().all(|ch| ch.is_ascii_digit()) {
-        return None;
-    }
-    Some((text[..dot + 2].to_string(), &text[dot + 2..]))
-}
-
-fn spans_plain_text(spans: &[Span]) -> String {
-    spans.iter().map(|span| span.content.as_ref()).collect()
-}
-
-fn digit_index(ch: char, len: usize) -> Option<usize> {
-    let digit = ch.to_digit(10)? as usize;
-    if digit == 0 || digit > len {
-        None
-    } else {
-        Some(digit - 1)
-    }
-}
-
-fn model_index_for_label(label: &str, options: &[ModelOption]) -> usize {
-    options
-        .iter()
-        .position(|option| option.spec.as_deref() == Some(label))
-        .unwrap_or(0)
-}
-
-fn custom_model_option() -> ModelOption {
-    ModelOption {
-        spec: None,
-        label: "Custom...".to_string(),
-        hint: "paste a model id".to_string(),
-    }
-}
-
-/// Convert models discovered from a provider's models API (already enriched
-/// with core profile names/descriptions) into picker options, keeping the
-/// trailing "Custom..." escape hatch.
-fn model_options_from_discovered(
-    models: Vec<crate::capabilities::model_discovery::DiscoveredProviderModel>,
-) -> Vec<ModelOption> {
-    /// OpenRouter descriptions run to paragraphs; the hint shares one row
-    /// with the model id, so keep it short.
-    const MAX_HINT_CHARS: usize = 72;
-
-    let mut options: Vec<ModelOption> = models
-        .into_iter()
-        .map(|model| {
-            let mut hint = match (model.display_name, model.description) {
-                (Some(name), Some(description)) => format!("{name} · {description}"),
-                (Some(name), None) => name,
-                (None, Some(description)) => description,
-                (None, None) => String::new(),
-            };
-            if hint.chars().count() > MAX_HINT_CHARS {
-                hint = hint.chars().take(MAX_HINT_CHARS - 1).collect::<String>() + "…";
-            }
-            ModelOption {
-                spec: Some(model.model_id.clone()),
-                label: model.model_id,
-                hint,
-            }
-        })
-        .collect();
-    options.push(custom_model_option());
-    options
-}
-
-/// Discovered model lists can be hundreds of entries (OpenRouter), far more
-/// than the setup panel can show. Window the list around the selection.
-const MAX_VISIBLE_MODEL_ROWS: usize = 8;
-
-fn model_window(selected: usize, total: usize, max_rows: usize) -> (usize, usize) {
-    if total <= max_rows {
-        return (0, total);
-    }
-    let start = selected.saturating_sub(max_rows / 2).min(total - max_rows);
-    (start, start + max_rows)
-}
-
-fn effort_index(value: &str) -> Option<usize> {
-    EFFORT_OPTIONS
-        .iter()
-        .position(|option| option.value.eq_ignore_ascii_case(value))
-}
-
-fn inset_x(area: Rect, pad: u16) -> Rect {
-    let total = pad.saturating_mul(2);
-    if area.width <= total {
-        return area;
-    }
-    Rect {
-        x: area.x.saturating_add(pad),
-        width: area.width.saturating_sub(total),
-        ..area
-    }
-}
-
-fn line_width(line: &Line) -> usize {
-    line.spans
-        .iter()
-        .map(|span| span.content.chars().count())
-        .sum()
-}
-
-fn separator_line(mut title: Line<'static>, width: u16, style: Style) -> Line<'static> {
-    let fill_width = (width as usize).saturating_sub(line_width(&title));
-    title
-        .spans
-        .push(Span::styled("─".repeat(fill_width), style));
-    title
-}
-
-fn draw_separator(f: &mut ratatui::Frame, area: Rect, title: Line<'static>, style: Style) {
-    if area.height == 0 {
-        return;
-    }
-    f.render_widget(
-        Paragraph::new(separator_line(title, area.width, style)),
-        area,
-    );
-}
-
-fn draw_input(f: &mut ratatui::Frame, area: Rect, app: &mut App) {
-    let area = inset_x(area, 0);
-    let prompt_width = area.width.min(2);
-    let prompt_area = Rect {
-        width: prompt_width,
-        ..area
-    };
-    let input_area = Rect {
-        x: area.x.saturating_add(prompt_width),
-        width: area.width.saturating_sub(prompt_width),
-        ..area
-    };
-    f.render_widget(
-        Paragraph::new(Span::styled(
-            "> ",
-            Style::default()
-                .fg(ACCENT_BLUE)
-                .add_modifier(Modifier::BOLD),
-        )),
-        prompt_area,
-    );
-    app.input.set_block(ratatui::widgets::Block::default());
-    f.render_widget(&app.input, input_area);
-    draw_input_cursor(f, input_area, app);
-}
-
-fn draw_input_cursor(f: &mut ratatui::Frame, area: Rect, app: &App) {
-    if app.busy || app.setup.is_some() {
-        return;
-    }
-
-    let inner_width = area.width;
-    let inner_height = area.height;
-    if inner_width == 0 || inner_height == 0 {
-        return;
-    }
-
-    let cursor = app.input.screen_cursor();
-    let x = area
-        .x
-        .saturating_add((cursor.col as u16).min(inner_width.saturating_sub(1)));
-    let y = area
-        .y
-        .saturating_add((cursor.row as u16).min(inner_height.saturating_sub(1)));
-    f.set_cursor_position((x, y));
-}
-
-fn message_separator_title(state: &ViewState) -> Line<'static> {
-    if state.busy {
-        return thinking_title(
-            state.busy_frame,
-            state.turn_activity.as_deref().unwrap_or("thinking"),
-        );
-    }
-    Line::from(vec![
-        Span::styled("─── ", Style::default().fg(ACCENT_BLUE)),
-        Span::styled(
-            format!("(Enter to send, {} for newline) ", newline_shortcut_hint()),
-            Style::default().fg(TEXT_MUTED),
-        ),
-    ])
-}
-
-fn newline_shortcut_hint() -> &'static str {
-    "Shift-Enter"
-}
-
-fn thinking_title(frame: u64, activity: &str) -> Line<'static> {
-    const SPINNER: [&str; 4] = ["-", "\\", "|", "/"];
-    let spinner = SPINNER[((frame / 2) as usize) % SPINNER.len()];
-    let text = format!("{activity}...");
-    let text_style = Style::default().fg(TEXT_MUTED).add_modifier(Modifier::BOLD);
-    let spans = vec![
-        Span::styled("─── ", Style::default().fg(ACCENT_BLUE)),
-        Span::styled(spinner.to_string(), Style::default().fg(ACCENT_GOLD)),
-        Span::raw(" "),
-        Span::styled(text, text_style),
-        Span::styled(" (input disabled) ", Style::default().fg(TEXT_DIM)),
-    ];
-    Line::from(spans)
-}
-
-fn draw_message_separator(f: &mut ratatui::Frame, area: Rect, state: &ViewState) {
-    draw_separator(
-        f,
-        area,
-        message_separator_title(state),
-        Style::default().fg(ACCENT_BLUE),
-    );
-}
-
-fn draw_status_separator(f: &mut ratatui::Frame, area: Rect) {
-    draw_separator(f, area, Line::from(""), Style::default().fg(ACCENT_GOLD));
-}
-
-fn draw_session_status(f: &mut ratatui::Frame, area: Rect, state: &ViewState) {
-    f.render_widget(
-        Paragraph::new(Line::from(vec![
-            Span::styled(" ", Style::default().fg(TEXT_MUTED)),
-            Span::styled(state.model_label.clone(), Style::default().fg(TEXT_MUTED)),
-            Span::styled("  ·  ", Style::default().fg(TEXT_DIM)),
-            Span::styled(
-                display_path(&state.workspace_root),
-                Style::default().fg(TEXT_MUTED),
-            ),
-            Span::styled("  ·  ", Style::default().fg(TEXT_DIM)),
-            Span::styled(
-                format!("{} msgs", state.lines_count),
-                Style::default().fg(TEXT_MUTED),
-            ),
-            Span::styled("  ·  approval ", Style::default().fg(TEXT_DIM)),
-            Span::styled(state.approval_mode.clone(), Style::default().fg(TEXT_MUTED)),
-            Span::styled("  ·  session ", Style::default().fg(TEXT_DIM)),
-            Span::styled(
-                state.session_id.to_string(),
-                Style::default().fg(TEXT_MUTED),
-            ),
-            Span::styled(" ", Style::default().fg(TEXT_MUTED)),
-        ])),
-        area,
-    );
-}
-
-fn display_path(path: &std::path::Path) -> String {
-    if let Ok(home) = std::env::var("HOME") {
-        let home = std::path::Path::new(&home);
-        if let Ok(rest) = path.strip_prefix(home) {
-            if rest.as_os_str().is_empty() {
-                return "~".to_string();
-            }
-            return format!("~/{}", rest.display());
-        }
-    }
-    path.display().to_string()
 }
 
 #[cfg(test)]
